requirements.txt
# agents-sdk-course-2/email-agent/requirements.txt
chainlit
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openpyxl # Required by pandas to read .xlsx files
google-generativeai # For Gemini LLM
pydantic # For data models like Email
